Move SearchResultAdapter to the common-ui module
