+++ feature-search/src/main/java/io/viewpoint/moviedatabase/ui/search/MovieSearchFragment.kt
@@ -13,6 +13,7 @@ import androidx.lifecycle.lifecycleScope
 import androidx.paging.LoadState
 import dagger.hilt.android.AndroidEntryPoint
 import io.viewpoint.moviedatabase.extensions.hideSoftInput
+import io.viewpoint.moviedatabase.ui.common.SearchResultAdapter
 import io.viewpoint.moviedatabase.ui.search.databinding.FragmentMovieSearchBinding
 import io.viewpoint.moviedatabase.ui.search.viewmodel.MovieSearchViewModel
 import kotlinx.coroutines.flow.collect
